docker/app_layer/onchain-stream-txs/requirements.txt
aiohttp==3.11.12
azure-identity==1.19.0
azure-keyvault==4.2.0
fastavro==1.10.0
confluent-kafka==2.6.0
hexbytes==1.3.0
kafka-python-ng==2.2.3
python-dotenv==0.16.0
redis==5.2.1
requests==2.28.1
web3==7.8.0